Renames instrument_summary to runs_list. Moves instrument:pause to insturment urls
