+++ WebApp/autoreduce_webapp/templates/instrument_variables.html
@@ -19,7 +19,7 @@
             <div class="row">
                 <div class="col-md-12 text-center">
                     <p>
-                        <a href="{% url 'instrument_summary' instrument=instrument.name %}">View upcoming saved configurations</a>
+                        <a href="{% url 'runs_list' instrument=instrument.name %}">View upcoming saved configurations</a>
                     </p>
                 </div>
             </div>
