+++ WebApp/autoreduce_webapp/templates/run_summary.html
@@ -67,7 +67,7 @@
                                 </div>
                                 <div>
                                     <strong>Instrument:</strong> <a
-                                        href="{% url 'instrument_summary' instrument=run.instrument.name %}">{{ run.instrument.name }}</a>
+                                        href="{% url 'runs_list' instrument=run.instrument.name %}">{{ run.instrument.name }}</a>
                                 </div>
                                 <div>
                                     <strong>RB Number:</strong> <a
